+++ draw_sphere.zig
@@ -38,8 +38,8 @@ pub fn main() !void {
     const origin = initPoint(0, 0, -5);
 
     const wall_z = 10.0;
-    const wall_size: f32 = 7.0;
-    const pixel_size = wall_size / @intToFloat(f32, CanvasSize);
+    const wall_size: f64 = 7.0;
+    const pixel_size = wall_size / @intToFloat(f64, CanvasSize);
     const half = 0.5 * wall_size;
 
     const light = PointLight{
@@ -49,13 +49,13 @@ pub fn main() !void {
 
     var y: usize = 0;
     while (y < canvas.height) : (y += 1) {
-        const world_y = half - @intToFloat(f32, y) * pixel_size;
+        const world_y = half - @intToFloat(f64, y) * pixel_size;
 
         var x: usize = 0;
         while (x < canvas.height) : (x += 1) {
             _ = 1;
 
-            const world_x = -half + @intToFloat(f32, x) * pixel_size;
+            const world_x = -half + @intToFloat(f64, x) * pixel_size;
             const wall_pos = initPoint(world_x, world_y, wall_z);
             const ray_dir = wall_pos.sub(origin).normalize();
             const ray = Ray.init(origin, ray_dir);
